+++ CPAC/longitudinal/wf/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020-2024  C-PAC Developers
+# Copyright (C) 2024  C-PAC Developers
 
 # This file is part of C-PAC.
 
@@ -14,7 +14,7 @@
 
 # You should have received a copy of the GNU Lesser General Public
 # License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
-"""Longitudinal preprocessing."""
+"""Workflows for longitudinal preprocessing."""
 
 from CPAC.utils.docs import DOCS_URL_PREFIX